+++ config/routes.php
@@ -1,8 +1,0 @@
-<?php
-
-use App\Http\Controllers\PageController;
-
-return [
-    'GET /' => [PageController::class, 'home'],
-    'GET /pages/{slug}' => [PageController::class, 'show'],
-];