+++ tests/Unit/ExampleTest.php
@@ -1,5 +1,0 @@
-<?php
-
-test('basic test', function () {
-    $this->assertTrue(true);
-});