+++ tests/root_tests/phpinfo.php
@@ -1,3 +1,0 @@
-<?php
-// Show PHP information
-phpinfo();