+++ tests/root_tests/test-page.php
@@ -1,3 +1,0 @@
-<?php
-// Simple test page
-phpinfo();